% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accel_nuclear_gradient_cpp}
\alias{accel_nuclear_gradient_cpp}
\title{Accelerated gradient function for ARD}
\usage{
accel_nuclear_gradient_cpp(
  inputs,
  outputs,
  lambda_val,
  Lipschitz = "regression",
  iterations = 5000L,
  etol = 1e-05,
  gamma = 2,
  symmetrized = TRUE,
  fixed_effects = FALSE
)
}
\arguments{
\item{inputs}{A matrix object. ARD census data (K x N).}

\item{outputs}{A matrix object. ARD survey data (K x M).}

\item{lambda_val}{A double. Initial lambda value.}

\item{Lipschitz}{String. Method for computing Lipschitz constant.}

\item{iterations}{Integer. Maximum number of iterations.}

\item{etol}{Double. Error tolerance.}

\item{gamma}{Double. Step size parameter.}

\item{symmetrized}{Boolean. Whether to symmetrize the output.}

\item{fixed_effects}{Boolean. Whether to use fixed effects.}
}
\value{
An N x M matrix estimate of network connections.
}
\description{
Accelerated gradient function for ARD
}
