% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation for ARD (R Implementation)}
\usage{
cross_validation(
  inputs,
  outputs,
  Lipschitz,
  iterations,
  etol,
  gamma,
  symmetrized,
  fixed_effects,
  CV_folds,
  method = c("grid", "mse"),
  lambda = NULL,
  CV_grid = NULL
)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.}

\item{outputs}{A matrix object. This contains the ARD survey data in matrix form.}

\item{Lipschitz}{A string. Method for computing the Lipschitz constant ("regression" or "JiYe").}

\item{iterations}{An integer. Maximum number of iterations.}

\item{etol}{A double. Error tolerance for convergence.}

\item{gamma}{A double. Step size parameter.}

\item{symmetrized}{A boolean. Whether to symmetrize the output.}

\item{fixed_effects}{A boolean. Whether to use fixed effects.}

\item{CV_folds}{An integer. Number of folds to use for cross-validation.}

\item{method}{A string. Determines the mode of cross validation.
Use "grid" to perform cross validation over a grid of lambda values or "mse" to compute the
cross validation mean squared error for a single lambda value.}

\item{lambda}{A double. A specific lambda value on which to compute the cross validation MSE.
Required if method = "mse".}

\item{CV_grid}{A NumericVector. Grid of lambda values to use for cross-validation.
Required if method = "grid".}
}
\value{
If method = "grid", returns a list containing:
\describe{
\item{best_lambda}{The lambda value corresponding to the minimum cross validation error.}
\item{min_error}{The minimum cross validation mean squared error.}
\item{var_errors}{The variance of the cross validation errors across folds.}
}
If method = "mse", returns a double representing the cross validation mean squared error for the provided lambda.
}
\description{
\code{cross_validation_combined} implements the cross validation algorithm outlined in Alidaee et al. (2020)
for selecting the optimal lambda value in ARD network estimation or computing mean squared error using cross validation.
}
