% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_validation_cpp}
\alias{cross_validation_cpp}
\title{Cross validation function for ARD}
\usage{
cross_validation_cpp(
  inputs,
  outputs,
  Lipschitz = "regression",
  iterations = 5000L,
  etol = 1e-05,
  gamma = 2,
  symmetrized = TRUE,
  fixed_effects = FALSE,
  CV_grid = as.numeric(c()),
  CV_folds = 5L
)
}
\arguments{
\item{inputs}{A matrix object. ARD census data (K x N).}

\item{outputs}{A matrix object. ARD survey data (K x M).}

\item{Lipschitz}{String. Method for computing Lipschitz constant.}

\item{iterations}{Integer. Maximum number of iterations.}

\item{etol}{Double. Error tolerance.}

\item{gamma}{Double. Step size parameter.}

\item{symmetrized}{Boolean. Whether to symmetrize the output.}

\item{fixed_effects}{Boolean. Whether to use fixed effects.}

\item{CV_grid}{NumericVector. Grid of lambda values to use for cross-validation. If not provided, defaults to a sequence from 0.01 to 10 with step size 0.01.}

\item{CV_folds}{Integer. Number of folds to use for cross-validation. Defaults to 5.}
}
\value{
A double. The optimal lambda value to use for network estimation.
}
\description{
Cross validation function for ARD
}
