% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{cross_validation_wrapper}
\alias{cross_validation_wrapper}
\title{Cross validation function for ARD (C++ wrapper)}
\usage{
cross_validation_wrapper(
  inputs,
  outputs,
  Lipschitz = "regression",
  iterations = 5000,
  etol = 1e-04,
  gamma = 2,
  symmetrize = TRUE,
  fixed_effects = FALSE,
  CV_grid = NULL,
  CV_folds = 5
)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.}

\item{outputs}{A matrix object. This contains the ARD survey data
in matrix form with size K x M, where N > M = number of households receiving ARD questionairre.}

\item{Lipschitz}{String. Method for computing Lipschitz constant.}

\item{iterations}{Integer. Maximum number of iterations.}

\item{etol}{Double. Error tolerance.}

\item{gamma}{Double. Step size parameter.}

\item{fixed_effects}{Boolean. Whether to use fixed effects.}

\item{CV_grid}{NumericVector. Grid of lambda values to use for cross-validation. If not provided, a dynamic grid search is conducted.}

\item{CV_folds}{Integer. Number of folds to use for cross-validation. Defaults to 5.}

\item{symmetrized}{Boolean. Whether to symmetrize the output.}
}
\value{
A double. The optimal lambda value to use for network estimation.
}
\description{
\code{cross_validation_wrapper} implements the cross validation algorithm outlined in Alidaee et al. (2020).
}
