% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{matrix_trace}
\alias{matrix_trace}
\title{Trace of a matrix}
\usage{
matrix_trace(mat)
}
\arguments{
\item{mat}{A matrix object. This can be a dcgMatrix from Matrix package. Must be have square dimensions.}
}
\value{
The trace of \code{mat}. This is a numeric, scalar value.
}
\description{
\code{matrix_trace} calculates the trace of
a single square (n x n) matrix.
}
\examples{
matrix_trace(matrix(1, nrow=3, ncol=3))

A <- matrix(2, nrow=4, ncol=4)
matrix_trace(A)

B <- Matrix::Matrix(1, nrow=3, ncol=3)
matrix_trace(B)

\dontrun{
A <- matrix(2, nrow=2, ncol=4)
matrix_trace(A)

matrix_trace(c(A, B))
matrix_trace(c(1,2,3))
}
}
\seealso{
More information on the trace of a matrix can be found at
\url{https://en.wikipedia.org/wiki/Trace_(linear_algebra)}.
}
