% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiYe_equations.R
\name{next_W_func}
\alias{next_W_func}
\title{Iterator for loss function}
\usage{
next_W_func(inputs, outputs, lambda, mu, W_k)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.
It should be of dimension K x N, where N = village size and K = number of ARD characteristics.}

\item{outputs}{A matrix object. This contains the ARD survey data
in matrix form with size K x M, where N > M = number of households receiving ARD questionairre.}

\item{lambda}{A scalar (numeric) value. It is provided by \code{\link{accel_nuclear_gradient}}.}

\item{mu}{A scalar (numeric) value. It is provided by \code{\link{accel_nuclear_gradient}}.}

\item{W_k}{A matrix object. This is the current iteration's guess for the ideal N x M linear operator.
Users do not need to create this themselves. It is provided by \code{\link{accel_nuclear_gradient}}
via \code{\link{next_W_func}}.}
}
\value{
A scalar value denoting the value of our approximate loss function.
}
\description{
\code{next_W_func} implements equation (15) from Ji & Ye (2009):
\deqn{W_k = p_L(W_{k - 1})} using the results from their Theorem 3.1,
which states that this can be found using singular value decomposition of \eqn{W_{k - 1}}.
}
\details{
This function should typically not be used directly. Instead, it should solely be called upon by other functions.
}
