% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{nuclear_norm}
\alias{nuclear_norm}
\title{Nuclear norm of a matrix}
\usage{
nuclear_norm(mat)
}
\arguments{
\item{mat}{A matrix object. This can be a dcgMatrix from Matrix package.}
}
\value{
The nuclear norm of \code{mat}. This is a numeric, scalar value.
}
\description{
\code{nuclear_norm} calculates the nuclear norm
for a single matrix. This is the Schatten norm in
the case of p = 1.
}
\details{
Because this is a nuclear norm, the function requires
input to be a matrix to avoid incorrect implementation.
}
\examples{
nuclear_norm(matrix(1, nrow=3, ncol=3))

A <- matrix(2, nrow=4, ncol=2)
nuclear_norm(A)

B <- Matrix::Matrix(1, nrow=3, ncol=3)
nuclear_norm(B)

\dontrun{
nuclear_norm(c(A, B))
nuclear_norm(c(1,2,3))
}
}
\seealso{
More information on the nuclear norm can be found at
\url{https://en.wikipedia.org/wiki/Matrix_norm#Schatten_norms}.
}
