% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclear_ard.R
\docType{package}
\name{nuclearARD}
\alias{nuclearARD-package}
\alias{nuclearARD}
\title{NuclearARD}
\description{
Package implementing nuclear norm regularization with aggregated relational data (ARD).
}
\author{
\strong{Maintainer}: Hossein Alidaee \email{halidaee@halidaee.com}

}
