% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiYe_equations.R
\name{obj_func.approx}
\alias{obj_func.approx}
\title{Approximated Loss function for ARD}
\usage{
obj_func.approx(inputs, outputs, W_k, W_kplus1, mu)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.
It should be of dimension K x N, where N = village size and K = number of ARD characteristics.}

\item{outputs}{A matrix object. This contains the ARD survey data
in matrix form with size K x M, where N > M = number of households receiving ARD questionairre.}

\item{W_k}{A matrix object. This is the current iteration's guess for the ideal N x M linear operator.
Users do not need to create this themselves. It is provided by \code{\link{accel_nuclear_gradient}}
via \code{\link{next_W_func}}.}

\item{W_kplus1}{A matrix object. This is the next iteration's guess for the ideal N x M linear operator.
Users do not need to create this themselves. It is provided by \code{\link{accel_nuclear_gradient}}
via \code{\link{next_W_func}}.}

\item{mu}{A scalar (numeric) value. It is provided by \code{\link{accel_nuclear_gradient}}.}
}
\value{
A scalar value denoting the value of our approximate loss function.
}
\description{
\code{obj_func.approx} implements equation (8) from Ji & Ye (2009):
\deqn{P_\mu (W, W_{k - 1}) + \lambda \Vert W \Vert_*.}{P_{\mu}(W, W_{k-1}) + \lambda ‖W‖_*.}
But---as in \code{\link{obj_func}} ---\eqn{\Vert \lambda \Vert_*}{\lambda ‖W‖_*} 
cancels out whenever needed, since the term is also in the objective function it is being compared to.
As a result, it is not necessary to compute. As a result, we are only actually only returning \eqn{P_\mu (W, W_{k - 1})}
This is defined by equation (7) of Ji & Ye: \deqn{P_\mu (W, W_{k - 1}) = f(W_{k - 1}) + 〈W - W_{k-1}, ∇f(W_{k-1}〉+ (\mu/2) \Vert W - W_{k-1}\Vert _F^2.}{P_\mu (W, W_{k - 1}) = f(W_{k - 1}) + 〈W - W_{k-1}, ∇f(W_{k-1}〉 + (\mu/2) ‖ W - W_{k-1}‖_F^2.}
}
\details{
This function should typically not be used directly. Instead, it should solely be called upon by other functions.
}
