% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiYe_equations.R
\name{obj_func}
\alias{obj_func}
\title{Objective (Loss) function for ARD}
\usage{
obj_func(inputs, outputs, W_k)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.
It should be of dimension K x N, where N = village size and K = number of ARD characteristics.}

\item{outputs}{A matrix object. This contains the ARD survey data
in matrix form with size K x M, where N > M = number of households receiving ARD questionairre.}

\item{W_k}{A matrix object. This is the current iteration's guess for the ideal N x M linear operator.
Users do not need to create this themselves. It is provided by \code{\link{accel_nuclear_gradient}}
via \code{\link{next_W_func}}.}
}
\value{
A scalar value denoting the value of our loss function.
}
\description{
\code{obj_func} calculates our objective function
\deqn{F(X) = \Vert  y - AX \Vert _2^2 + \lambda \Vert  X \Vert _*}{F(X) = ‖y - AX‖_2^2 + \lambda ‖X‖_*}
but---as stated in \code{\link{obj_func.approx}} ---\eqn{\Vert \lambda \Vert_*}{\lambda ‖X‖_*} 
cancels out whenever needed, since the term is also in the approximated function it is being compared to.
As a result, it is not necessary to compute. Because of this, we effectively
are calculating \eqn{f} instead of \eqn{F}.
}
\details{
This function should typically not be used directly. Instead, it should solely be called upon by other functions.
}
