% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{obj_func.testing}
\alias{obj_func.testing}
\title{Input validation for objective function}
\usage{
obj_func.testing(inputs, outputs, W_k)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.
It should be of dimension K x N, where N = village size and K = number of ARD characteristics.}

\item{outputs}{A matrix object. This contains the ARD survey data
in matrix form with size K x M, where N > M = number of households receiving ARD questionairre.}

\item{W_k}{A matrix object. This is the current iteration's guess for the ideal N x M linear operator.
Users do not need to create this themselves. It is provided via \code{\link{next_W_func}}.}
}
\value{
There is no returned value. This function solely raises errors in the event of invalid inputs.
}
\description{
\code{\link{obj_func}} and \code{\link{obj_func.grad}} have common inputs
and therefore require the same validations. To avoid code redundancy, these
common checks have been placed in this function, which is then referenced.
}
