% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrize a matrix}
\usage{
symmetrize(mat)
}
\arguments{
\item{mat}{A matrix object. This can be a dcgMatrix from Matrix package. Must be have square dimensions.}
}
\value{
The symmetric transfomation of \code{mat}.
}
\description{
\code{symmetrize} transforms an m x n matrix, where m > n, to be
symmetric in the top n x n submatrix, with zero diagonals.
}
\examples{
symmetrize(matrix(1, nrow=3, ncol=3))

A <- matrix(2, nrow=6, ncol=4)
symmetrize(A)

\dontrun{
A <- matrix(2, nrow=2, ncol=4)
symmetrize(A)
}
}
