% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation for ARD (R Implementation)}
\usage{
cross_validation(
  inputs,
  outputs,
  Lipschitz,
  iterations,
  etol,
  gamma,
  symmetrized,
  fixed_effects,
  CV_grid,
  CV_folds
)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.}

\item{outputs}{A matrix object. This contains the ARD survey data in matrix form.}

\item{Lipschitz}{A string. Method for computing the Lipschitz constant ("regression" or "JiYe").}

\item{iterations}{An integer. Maximum number of iterations.}

\item{etol}{A double. Error tolerance for convergence.}

\item{gamma}{A double. Step size parameter.}

\item{symmetrized}{A boolean. Whether to symmetrize the output.}

\item{fixed_effects}{A boolean. Whether to use fixed effects.}

\item{CV_grid}{A NumericVector. Grid of lambda values to use for cross-validation.}

\item{CV_folds}{An integer. Number of folds to use for cross-validation.}
}
\value{
A double. The optimal lambda value to use for network estimation.
}
\description{
\code{cross_validation} implements the cross validation algorithm outlined in Alidaee et al. (2020)
for selecting the optimal lambda value in ARD network estimation.
}
