% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_accelerated_descent.R
\name{matrix_regression}
\alias{matrix_regression}
\title{Matrix regression function for ARD}
\usage{
matrix_regression(inputs, outputs, iterations = 5000, etol = 1e-04)
}
\arguments{
\item{inputs}{A matrix object. This contains ARD census data in matrix form.
It should be of dimension K x N, where N = village size and K = number of ARD characteristics.}

\item{outputs}{A matrix object. This contains the ARD survey data
in matrix form with size K x M, where N > M = number of households receiving ARD questionairre.}

\item{iterations}{A scalar (integer) value. It is the number of iterations the user
specifies should occur. Set by default to 5000.}

\item{etol}{A scaler. It is the error tolerance for the algorithm. The algorithm will terminate
when either the maximum number of iterations has been met or the mean absolute error between iterations
is below etol.}
}
\value{
An N x M matrix estimate of network connections.
}
\description{
\code{matrix_regression} is a wrapper for \code{\link{accel_nuclear_gradient}} for the specific
implementation of the multivariate regression discussed in Alidaee, Auerbach, and Leung (2020).
Consequently, options have been set to the optimal values of accelerated gradient descent
in the case of a multivariate regression.
}
